# Updated variableScope.has() method to search in layers