--- a/lib/collection/variable-scope.js
+++ b/lib/collection/variable-scope.js
@@ -197,7 +197,17 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
      * @returns {Boolean}
      */
     has: function (variableName) {
-        var variable = this.values.oneNormalizedVariable(variableName);
+        var variable = this.values.oneNormalizedVariable(variableName)
+            i,
+            ii;
+
+            // if a variable does not exist in local scope, we search all the layers and return the first occurrence.
+        if (!(variable || !this._layers)) {
+            for (i = 0, ii = this._layers.length; i < ii; i++) {
+                variable = this._layers[i].oneNormalizedVariable(key);
+                if (variable && !variable.disabled) { break; }
+            }
+        }
 
         return Boolean(variable && !variable.disabled);
     },