--- a/lib/collection/variable-scope.js
+++ b/lib/collection/variable-scope.js
@@ -193,15 +193,15 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
     /**
      * Determines whether one particular variable is defined in this scope of variables.
      *
-     * @param {String} variableName
+     * @param {String} key
      * @returns {Boolean}
      */
-    has: function (variableName) {
-        var variable = this.values.oneNormalizedVariable(variableName)
+    has: function (key) {
+        var variable = this.values.oneNormalizedVariable(key),
             i,
             ii;
 
-            // if a variable does not exist in local scope, we search all the layers and return the first occurrence.
+        // if a variable does not exist in local scope, we search all the layers and return the first occurrence.
         if (!(variable || !this._layers)) {
             for (i = 0, ii = this._layers.length; i < ii; i++) {
                 variable = this._layers[i].oneNormalizedVariable(key);