--- a/lib/collection/variable-scope.js
+++ b/lib/collection/variable-scope.js
@@ -193,7 +193,7 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
     /**
      * Determines whether one particular variable is defined in this scope of variables.
      *
-     * @param {String} key
+     * @param {String} key - The name of the variable to check
      * @returns {Boolean}
      */
     has: function (key) {
@@ -201,8 +201,9 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
             i,
             ii;
 
-        // if a variable does not exist in local scope, we search all the layers and return the first occurrence.
-        if (!(variable || !this._layers)) {
+        // if a variable is disabled or does not exist in local scope,
+        // we search all the layers and return the first occurrence.
+        if ((!variable || variable.disabled) && this._layers) {
             for (i = 0, ii = this._layers.length; i < ii; i++) {
                 variable = this._layers[i].oneNormalizedVariable(key);
                 if (variable && !variable.disabled) { break; }