--- a/lib/collection/variable-scope.js
+++ b/lib/collection/variable-scope.js
@@ -194,7 +194,8 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
      * Determines whether one particular variable is defined in this scope of variables.
      *
      * @param {String} key - The name of the variable to check
-     * @returns {Boolean} - True if variable with given key is present and enabled in any layers, false otherwise
+     * @returns {Boolean} - Returns true if an enabled variable with given key is present in current or parent scopes,
+     *                      false otherwise
      */
     has: function (key) {
         var variable = this.values.oneNormalizedVariable(key),