--- a/lib/collection/variable-scope.js
+++ b/lib/collection/variable-scope.js
@@ -204,14 +204,14 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
 
         // if a variable is disabled or does not exist in local scope,
         // we search all the layers and return the first occurrence.
-        if ((!variable || variable.disabled) && this._layers) {
+        if ((!variable || variable.disabled === true) && this._layers) {
             for (i = 0, ii = this._layers.length; i < ii; i++) {
                 variable = this._layers[i].oneNormalizedVariable(key);
                 if (variable && !variable.disabled) { break; }
             }
         }
 
-        return Boolean(variable && !variable.disabled);
+        return Boolean(variable && variable.disabled !== true);
     },
 
     /**