--- a/lib/collection/variable-scope.js
+++ b/lib/collection/variable-scope.js
@@ -207,7 +207,7 @@ _.assign(VariableScope.prototype, /** @lends VariableScope.prototype */ {
         if ((!variable || variable.disabled === true) && this._layers) {
             for (i = 0, ii = this._layers.length; i < ii; i++) {
                 variable = this._layers[i].oneNormalizedVariable(key);
-                if (variable && !variable.disabled) { break; }
+                if (variable && variable.disabled !== true) { break; }
             }
         }