Updated variableScope.has() method to search given variable in layers of the scope.

Solves: https://github.com/postmanlabs/postman-app-support/issues/6595

Benchmarks before update:
```
Search variable in first layer:
    5 layers of 10 variables: 14,699,406 ops/sec (±1.23%, ⨉1000000)
    5 layers of 100 variables: 9,606,022 ops/sec (±4.52%, ⨉1000000)
    5 layers of 1000 variables: 10,110,274 ops/sec (±0.58%, ⨉1000000)

Search variable in last layer:
    5 layers of 10 variables: not supported
    5 layers of 100 variables: not supported
    5 layers of 1000 variables: not supported
```

Benchmarks after update:
```
Search variable in first layer:
    5 layers of 10 variables: 14,180,963 ops/sec (±0.52%, ⨉1000000)
    5 layers of 100 variables: 10,084,397 ops/sec (±1.98%, ⨉1000000)
    5 layers of 1000 variables: 10,542,299 ops/sec (±1.02%, ⨉1000000)

Search variable in last layer:
    5 layers of 10 variables: 3,142,960 ops/sec (±0.97%, ⨉1000000)
    5 layers of 100 variables: 790,895 ops/sec (±0.31%, ⨉1000000)
    5 layers of 1000 variables: 86,119 ops/sec (±0.14%, ⨉430598)
```